add BaseDelimitedTextOperation to encapsulated common logic
