+++ ContentPatcher/Framework/ConfigModels/TextOperationConfig.cs
@@ -24,16 +24,16 @@ namespace ContentPatcher.Framework.ConfigModels
         /****
         ** Append/Prepend/ReplaceDelimited
         ****/
-        /// <summary>The value to append, prepend, or replace with.</summary>
+        /// <summary>The operation value (e.g. the value to append or replace with).</summary>
         public string? Value { get; }
 
         /****
         ** RemoveDelimited/ReplaceDelimited
         ****/
-        /// <summary>The value to remove from or replace in the text.</summary>
+        /// <summary>The value to match in the text.</summary>
         public string? Search { get; }
 
-        /// <summary>Which delimited values should be removed or replaced.</summary>
+        /// <summary>Which delimited values should be matched.</summary>
         public string? ReplaceMode { get; }
 
 
@@ -43,10 +43,10 @@ namespace ContentPatcher.Framework.ConfigModels
         /// <summary>Construct an instance.</summary>
         /// <param name="operation">The text operation to perform.</param>
         /// <param name="target">The specific text field to change as a breadcrumb path. Each value in the list represents a field to navigate into.</param>
-        /// <param name="value">The value to append, prepend, or replace with.</param>
+        /// <param name="value">The operation value (e.g. the value to append or replace with).</param>
         /// <param name="delimiter">If the target field already has a value, text to add between the previous and inserted values, if any.</param>
-        /// <param name="search">The value to remove from or replace in the text.</param>
-        /// <param name="replaceMode">Which delimited values should be removed or replaced.</param>
+        /// <param name="search">The value to match in the text.</param>
+        /// <param name="replaceMode">Which delimited values should be matched.</param>
         [JsonConstructor]
         public TextOperationConfig(string? operation, string?[]? target, string? value, string? delimiter, string? search, string? replaceMode)
         {
